src/app/modules/common-components/snackbar/snackbar.component.spec.ts
import { DebugElement } from '@angular/core';
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { of } from 'rxjs';
import { SnackbarComponent } from './snackbar.component';

describe('SnackbarComponent', () => {
  let component: SnackbarComponent;
  let fixture: ComponentFixture<SnackbarComponent>;
  let debugElement: DebugElement;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [BrowserAnimationsModule],
      declarations: [SnackbarComponent],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(SnackbarComponent);
    component = fixture.componentInstance;
    debugElement = fixture.debugElement;
    component.snackbarLabel$ = of('label');
    spyOn(component, 'dismissSnackbar');
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should set the label and then dismiss the snackbar on getting a label', () => {
    expect(component.label).toEqual('label');
    expect(component.shouldShowSnackbar).toBeTruthy();
    expect(component.dismissSnackbar).toHaveBeenCalled();
  });
});
